# ticketing/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the TicketPulse API for ticketing microservice. Concerts, sessions, sections and tickets are all here.",
        "title": "TicketPulse Ticketing API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/sections": {
            "post": {
                "description": "Create a new section and its associated tickets",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Create a section with tickets",
                "parameters": [
                    {
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Section Name",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Section Capacity",
                        "name": "capacity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Section Category",
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Ticket Price",
                        "name": "price",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/sections/{id}": {
            "get": {
                "description": "Retrieve details of a specific section by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get details of a section",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Section ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.SectionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update details of a specific section by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Update a section",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Section ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Section Name",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Section Capacity",
                        "name": "capacity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Reserved Seats",
                        "name": "reserved",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Bought Seats",
                        "name": "bought",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Section Category",
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Ticket Price",
                        "name": "price",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.SectionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.SuccessResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a specific section by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Delete a section",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Section ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/admin/sections/{id}/tickets": {
            "get": {
                "description": "Retrieve all tickets associated with a specific section by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get tickets by section",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Section ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.SectionWithTicketsListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.SuccessResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ticketing_internal_common_types.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "ticketing_internal_common_types.Section": {
            "type": "object",
            "properties": {
                "bought": {
                    "type": "integer"
                },
                "capacity": {
                    "type": "integer"
                },
                "category": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "reserved": {
                    "type": "integer"
                }
            }
        },
        "ticketing_internal_common_types.SectionResponse": {
            "type": "object",
            "properties": {
                "section": {
                    "$ref": "#/definitions/ticketing_internal_common_types.Section"
                }
            }
        },
        "ticketing_internal_common_types.SectionWithTicketsListResponse": {
            "type": "object",
            "properties": {
                "section": {
                    "type": "string"
                },
                "tickets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ticketing_internal_common_types.Ticket"
                    }
                }
            }
        },
...
        "ticketing_internal_common_types.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "ticketing_internal_common_types.Ticket": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "seatNumber": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        }
    }
}